Updates to exec summary regarding relative fishing intensity
Note that we use relative fishing intenstity for almost everything
(i.e., because relative the target is 1 or 100%), except for we use just
plain-o fishing intensity in tables i and j and figures j and k because
we are referring to the SPR fishing intensity target directly (not in a
relative sense) where the target is 40% (not 100% as in relative case).
All this is similar to the way others think about biomass targets right,
SB40% could be a target so your refer to the 40% level, or SB40%/SB0
could be a target such that 1 (or 100%) is equating to the target.
Confusing I know, but I think it bit better now...