Add table of model changes for SRG request, part of #463
Table includes information since 2008, but only since 2011 is displayed
because @taylori mentioned that prior to 2012 multiple modelling frameworks
were used and it would be unfair to mainly document changes in SS.
Table will need to be manually updated each year. Could automate some of it
in the future, but not right now.
More text will come in a subsequent commit about data-weighting.