Add functionality to make.risk.table for additional column of catches
Must use type = 2, this shrinks some of the column names and sets the
column widths to be a fixed rather than dynamic value. The default is
to have the table with only one column of catches as has been done in
previous years. type = 2 will only work if there are two years of
projections in the risk table, which is currently true. #436