Add plotting code for fishing and bottom depths
Also added plots to the data presentation but they are
currently commented out. These plots are in m rather than fathoms
so I will wait to see if there is a matching canadian csv that
can be plotted before switching the figure. Will need to change
the assessment document as well. For #566