Commits relevant to #539 1998 survey year wrong in dat file
The 1998 survey estimate and se were from the 2016 model file
instead of the revised 2017 estimate. The same problem was in
the 2018 dat file but was not found until 2019.
Some information still needs to be added to the assessment document,
upon discussion no additional documentation is needed.

Additionally, Table 14 was removed b/c it duplicated Table 13 and
Figure 9 was removed because the only difference between
it and Figure 10 was the variability due to squid, which was added
to Figure 10 by including "4" in the show argument.