feature: description file to list packages needed to build document
The Description file will be a long-standing method of listing necessary
packages to build the hake assessment. I think we can keep going forward
with listing new packages that people need to install as an issue
but then also list them in the DESCRIPTION file so that many months
from the change or for new people, it will be easy to know what
packages are required.