fix: development branch vs master of r4ss
(1) doesn't make comment in wtatage, fix assumes
  (a) fecundity is fleet -2 and
  (b) mid-term population is fleet 1.

(2) Catch_or_F in forecast is now 'Catch or F', which was leading to the
forecasts not running, use grep to find the right information.

(3) plotting of mcmc pars (M, R_0, h, and extra CV) uses SSplotPars, a
function that has had arguments removed from its call, one of them being
'new' which was set to FALSE, i.e., not the default value of new = TRUE,
so I couldn't remove it otherwise the code woudln't work for the master
branch. Using do.call I add the argument to the list used to call SSplotPars.