close #591 add reference to Stienessen in Research and Data needs
JMC wanted a note about the use of Vessels-of-Opportunity in the
assessment, but I found that it was already in the Research and
Data Needs. I added a reference to the Pollock fishery where they
use the fishery-dependent data as a index of abundance in their
assessment b/c their survey is not operated annual. This index is
compiled from acoustic information that fishing vessels return to
researchers.

Also, fishing vessels provided temperature data which were then analyzed
against bycatch to see if there was a link, which there was. Then
this temperature information was archived and helped inform general
knowledge about the ecosystem.