Work on #631 to automate heat map
* fixed the last year of the model issue to be the last year of the
data now that SS can use the last year of data as empirical.

* added a new line of zero values for no interpolation in 2019,
next year fill this in from the matrix of sample sizes

* Allow for longterm.mean.ages to be NULL, which will take the first
year of the matrix and assume that it is the mean.

* This would all be better if it were plotted using the raw data
rather than the matrix input into SS.
pacific-hake/hakedataUSA#5 will move the code for the 2021 assessment