remove posterior.regex from make.short.parameter.estimates.sens.table
because the table can't be built with different parameters, the names
in the output table won't line up. I forgot to remove the argument
when I was working with the table yesterday.

Next I will work with the other table that is like this one trying
to reduce the amount of code that needs to be maintained.