update and fix US Catches (part of #708)
* add 2020 Atsea catches to us-catch-by-month
* update landings-tac-history with code to bring in US csv files and sum them
* digits might still be off, or not standardized but i wanted to get this committed

MAJOR NOTE:
* fix historical landings for shoreside, lots of time went into trying
to determine how long and why there were differences
* shoreside is coming straight out of PacFIN, and since 2016 (the last
file found) the shoreside catches in 1986 have been 3431 not 3463 or 3465.
So, I am not sure when PacFIN changed but the assessments list
assessmentyear mt
1991 3463
2010 3.46x1000
2011 3.47x1000
2014 3465
2020 3431.947
so this year has been off by 33.4 mt for a while. The differences are from
July of that year, but I don't know any more than that b/c data pulls
from PacFIN are not available prior to 2016.