"fishery selectivity parameters" when talking about time varying
close #377
I double checked the control file, we only have time-varying selectivity
for the fishery in years 1991-present. The p_a values are for
both types of selectivity, but we only have additional phi parameters
for the fishery. So the document is correct with the equations.

I removed "Note that there is such a loglikelihood ..." because there
is a loglikelihood for both but phi is zero for the survey, so it
was confusing.