Uses new files us- and survey-weight.csv
It was really confusing to know what had been and what had not been
updated, so now all weight-at-age data are in one of three summary files.

The Canadian data contained some duplicates so @cgrandin provided a new
file, and thus, this commit removes the old one to reduce confusion.