Adds mean weight-at-age as a listed data source
Mean weight-at-age was not listed in the text as a data source yet
it is used. I tried to turn on mean weight-at-age data in the data
plot but it is not doable as of right now. We might think about adding
that to r4ss. @iantaylor-NOAA has anyone ever suggested that it be
included? Seems like you could skip all years before equilibrium and
interpolated years i.e., those that are filled in from years with a
negative year. Happy to make an issue in r4ss if you think it would
be a good addition. Otherwise we can just go on as is.
Close #971