Use \rightmark to define the right footer
Wanted to remove links, i.e., \ref{}, from the right footer
because all links are tagged, which made the footer have tagging.
Paragraphs are now better tagged because they often span pages
and the link is no longer in the middle of them.

This could be made better with more use of \fancyhdr and \makeright
but it is slightly convoluted because we have a mix of chapter, numbered
sections, and un-numbered sections that we want to be part of the
right-footer contents. AND some chapter names are REALLY long and end
up covering the page number in the center.