fix(create_sens_dirs): Remove meaningless messages
We do not need the exe because it is in the path. So, remove the object
and the message. Use `fs::dir_create()` rather than `dir.create()`
because the latter gives you a warning if it already exists. And, get
rid of relic reading of the starter file because `SS_read()` does it
all for you.