fix: Changes 2024 to 2023 for Aaron's presentation
on dynamic b_0 in the SRG responses, also fixes a typo that I made
but I cannot remember what word it was.
Re-wrapped responses at 80 characters and added greater than signs to
the responses that were missing them to allow for the automatic
indentation. Sorry should have broke this up into a couple of commits.